+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/reference/cis.ts
@@ -1,0 +1,4 @@
+export interface CIS {
+	code: string // CIS source code
+	name: string // Human-readable name
+}
