+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/reference/loading-category.ts
@@ -1,0 +1,14 @@
+export interface LoadingCategory {
+	$: {
+		Code: string // Loading category code
+		Name: string // Human-readable message
+		Toc?: string // TOC (operator) code
+	}
+	$$: {
+		typicaldescription: string[]
+		expecteddescription: string[]
+		definition: string[]
+		colour?: string[]
+		image?: string[]
+	}
+}
