+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/reference/reason.ts
@@ -1,0 +1,4 @@
+export interface Reason {
+	code: string // Reason code
+	reasontext: string // Human-readable reason
+}
