+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/reference/toc.ts
@@ -1,0 +1,5 @@
+export interface TrainOperatingCompany {
+	toc: string // TOC (operator) code
+	tocname: string // Human-readable name
+	url: string // Map URL
+}
