+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/reference/via.ts
@@ -1,0 +1,9 @@
+export interface Via {
+	at: string // TIPLOC
+	dest: string // TIPLOC
+
+	loc1: string // TIPLOC
+	loc2?: string // TIPLOC
+
+	viatext: string // Human-readable announcer message
+}
