+++ source/server/sources/national-rail-data-portal/darwin-push-port/types/time-table/time-table.ts
@@ -1,0 +1,12 @@
+import { Journey } from "./journey.js"
+
+export interface TimeTable {
+	pporttimetable: {
+		$: {
+			timetableID: string
+		}
+		$$: {
+			journey: Journey[]
+		}
+	}
+}
