+++ source/server/state.ts
@@ -1,0 +1,5 @@
+import { TimeTable } from "./classes/timetable.js"
+
+let timeTable: TimeTable | null = null
+export const setTimeTable = (value: TimeTable): TimeTable => (timeTable = value)
+export const getTimeTable = (): TimeTable | null => timeTable
